Move functions to utils.go to allow testing
