+++ longtailstorelib/blobStore_test.go
@@ -6,6 +6,8 @@ import (
 	"strings"
 	"sync"
 	"testing"
+
+	"github.com/DanEngelbrecht/golongtail/longtaillib"
 )
 
 type testBlob struct {
@@ -106,6 +108,74 @@ func (blobObject *testBlobObject) Delete() error {
 
 	delete(blobObject.client.store.blobs, blobObject.path)
 	return nil
+}
+
+func generateStoredBlock(t *testing.T, seed uint8) (longtaillib.Longtail_StoredBlock, int) {
+	chunkHashes := []uint64{uint64(seed) + 1, uint64(seed) + 2, uint64(seed) + 3}
+	chunkSizes := []uint32{uint32(seed) + 10, uint32(seed) + 20, uint32(seed) + 30}
+
+	blockDataLen := (int)(chunkSizes[0] + chunkSizes[1] + chunkSizes[2])
+	blockData := make([]uint8, blockDataLen)
+	for p := 0; p < blockDataLen; p++ {
+		blockData[p] = seed
+	}
+
+	return longtaillib.CreateStoredBlock(
+		uint64(seed)+21412151,
+		997,
+		2,
+		chunkHashes,
+		chunkSizes,
+		blockData,
+		false)
+}
+
+func storeBlockFromSeed(t *testing.T, storeAPI longtaillib.Longtail_BlockStoreAPI, seed uint8) (uint64, int) {
+	storedBlock, errno := generateStoredBlock(t, seed)
+	if errno != 0 {
+		return 0, errno
+	}
+
+	p := &putStoredBlockCompletionAPI{}
+	p.wg.Add(1)
+	errno = storeAPI.PutStoredBlock(storedBlock, longtaillib.CreateAsyncPutStoredBlockAPI(p))
+	if errno != 0 {
+		p.wg.Done()
+		storedBlock.Dispose()
+		return 0, errno
+	}
+	p.wg.Wait()
+
+	return uint64(seed) + 21412151, p.err
+}
+
+func fetchBlockFromStore(t *testing.T, storeAPI longtaillib.Longtail_BlockStoreAPI, blockHash uint64) (longtaillib.Longtail_StoredBlock, int) {
+	g := &getStoredBlockCompletionAPI{}
+	g.wg.Add(1)
+	errno := storeAPI.GetStoredBlock(blockHash, longtaillib.CreateAsyncGetStoredBlockAPI(g))
+	if errno != 0 {
+		g.wg.Done()
+		return longtaillib.Longtail_StoredBlock{}, errno
+	}
+	g.wg.Wait()
+
+	return g.storedBlock, g.err
+}
+
+func validateBlockFromSeed(t *testing.T, seed uint8, storedBlock longtaillib.Longtail_StoredBlock) {
+	if !storedBlock.IsValid() {
+		t.Errorf("validateBlockFromSeed() g.err %t != %t", storedBlock.IsValid(), true)
+	}
+
+	storedBlockIndex := storedBlock.GetBlockIndex()
+
+	if storedBlockIndex.GetBlockHash() != uint64(seed)+21412151 {
+		t.Errorf("validateBlockFromSeed() storedBlockIndex.GetBlockHash() %d != %d", storedBlockIndex.GetBlockHash(), uint64(seed)+21412151)
+	}
+
+	if storedBlockIndex.GetChunkCount() != 3 {
+		t.Errorf("validateBlockFromSeed() storedBlockIndex.GetChunkCount() %d != %d", storedBlockIndex.GetChunkCount(), 3)
+	}
 }
 
 func TestCreateStoreAndClient(t *testing.T) {
