+++ longtailstorelib/remotestore.go
@@ -2,7 +2,6 @@ package longtailstorelib
 
 import (
 	"context"
-	"crypto/sha1"
 	"fmt"
 	"log"
 	"path/filepath"
@@ -96,48 +95,6 @@ func (s *remoteStore) String() string {
 	return s.defaultClient.String()
 }
 
-func readBlobWithRetry(
-	ctx context.Context,
-	client BlobClient,
-	key string) ([]byte, int, error) {
-	retryCount := 0
-	objHandle, err := client.NewObject(key)
-	if err != nil {
-		return nil, retryCount, err
-	}
-	exists, err := objHandle.Exists()
-	if err != nil {
-		return nil, retryCount, err
-	}
-	if !exists {
-		return nil, retryCount, longtaillib.ErrENOENT
-	}
-	blobData, err := objHandle.Read()
-	if err != nil {
-		log.Printf("Retrying getBlob %s using %s\n", key, client.String())
-		retryCount++
-		blobData, err = objHandle.Read()
-	}
-	if err != nil {
-		log.Printf("Retrying 500 ms delayed getBlob %s using %s\n", key, client.String())
-		time.Sleep(500 * time.Millisecond)
-		retryCount++
-		blobData, err = objHandle.Read()
-	}
-	if err != nil {
-		log.Printf("Retrying 2 s delayed getBlob %s using %s\n", key, client.String())
-		time.Sleep(2 * time.Second)
-		retryCount++
-		blobData, err = objHandle.Read()
-	}
-
-	if err != nil {
-		return nil, retryCount, err
-	}
-
-	return blobData, retryCount, nil
-}
-
 func putStoredBlock(
 	ctx context.Context,
 	s *remoteStore,
@@ -150,43 +107,20 @@ func putStoredBlock(
 	blockIndex := storedBlock.GetBlockIndex()
 	blockHash := blockIndex.GetBlockHash()
 	key := GetBlockPath("chunks", blockHash)
-	objHandle, err := blobClient.NewObject(key)
+	blob, errno := longtaillib.WriteStoredBlockToBuffer(storedBlock)
+	if errno != 0 {
+		return longtaillib.ErrnoToError(errno, longtaillib.ErrEIO)
+	}
+
+	retryCount, err := writeBlobWithRetry(ctx, blobClient, key, blob)
+	atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_RetryCount], uint64(retryCount))
 	if err != nil {
+		atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_FailCount], 1)
 		return err
 	}
-	if exists, err := objHandle.Exists(); err == nil && !exists {
-		blob, errno := longtaillib.WriteStoredBlockToBuffer(storedBlock)
-		if errno != 0 {
-			return longtaillib.ErrnoToError(errno, longtaillib.ErrEIO)
-		}
 
-		ok, err := objHandle.Write(blob)
-		if err != nil || !ok {
-			log.Printf("Retrying putBlob %s in store %s\n", key, s.String())
-			atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_RetryCount], 1)
-			ok, err = objHandle.Write(blob)
-		}
-		if err != nil || !ok {
-			log.Printf("Retrying 500 ms delayed putBlob %s in store %s\n", key, s.String())
-			time.Sleep(500 * time.Millisecond)
-			atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_RetryCount], 1)
-			ok, err = objHandle.Write(blob)
-		}
-		if err != nil || !ok {
-			log.Printf("Retrying 2 s delayed putBlob %s in store %s\n", key, s.String())
-			time.Sleep(2 * time.Second)
-			atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_RetryCount], 1)
-			ok, err = objHandle.Write(blob)
-		}
-
-		if err != nil || !ok {
-			atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_FailCount], 1)
-			return longtaillib.ErrnoToError(errno, longtaillib.ErrEIO)
-		}
-
-		atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_Byte_Count], (uint64)(len(blob)))
-		atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_Chunk_Count], (uint64)(blockIndex.GetChunkCount()))
-	}
+	atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_Byte_Count], (uint64)(len(blob)))
+	atomic.AddUint64(&s.stats.StatU64[longtaillib.Longtail_BlockStoreAPI_StatU64_PutStoredBlock_Chunk_Count], (uint64)(blockIndex.GetChunkCount()))
 
 	// We need to make a copy of the block index - as soon as we call the OnComplete function the data for the block might be deallocated
 	storedBlockIndex, errno := longtaillib.WriteBlockIndexToBuffer(blockIndex)
@@ -422,156 +356,6 @@ func remoteWorker(
 	return nil
 }
 
-func updateRemoteStoreIndex(
-	ctx context.Context,
-	blobClient BlobClient,
-	addedStoreIndex longtaillib.Longtail_StoreIndex) error {
-
-	addedStoreIndexBlob, errno := longtaillib.WriteStoreIndexToBuffer(addedStoreIndex)
-	if errno != 0 {
-		return longtaillib.ErrnoToError(errno, longtaillib.ErrENOMEM)
-	}
-	addedStoreIndexSha := sha1.Sum(addedStoreIndexBlob)
-	addedStoreIndexName := fmt.Sprintf("index/%x.plsi", addedStoreIndexSha)
-	objHandle, err := blobClient.NewObject(addedStoreIndexName)
-	ok, err := objHandle.Write(addedStoreIndexBlob)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		// Return error here
-		return nil
-	}
-
-	storeIndex, partialIndexNames, err := readStoreIndexes(ctx, blobClient)
-	defer storeIndex.Dispose()
-	if err != nil {
-		return err
-	}
-
-	storeIndexBlob, errno := longtaillib.WriteStoreIndexToBuffer(storeIndex)
-	objHandle, err = blobClient.NewObject("store.lsi")
-	ok, err = objHandle.Write(storeIndexBlob)
-	if err != nil {
-		return err
-	}
-	if !ok {
-		// Return error here
-		return nil
-	}
-
-	// Loop hole here? Need to think hard if it is possible to mess things up...
-	for _, indexName := range partialIndexNames {
-		objHandle, err := blobClient.NewObject(indexName)
-		if err == nil {
-			objHandle.Delete()
-		}
-	}
-	return nil
-}
-
-func getStoreIndexFromBlocks(
-	ctx context.Context,
-	s *remoteStore,
-	blobClient BlobClient,
-	blockKeys []string) (longtaillib.Longtail_StoreIndex, error) {
-
-	storeIndex, errno := longtaillib.CreateStoreIndexFromBlocks([]longtaillib.Longtail_BlockIndex{})
-	if errno != 0 {
-		return longtaillib.Longtail_StoreIndex{}, longtaillib.ErrnoToError(errno, longtaillib.ErrENOMEM)
-	}
-
-	batchCount := runtime.NumCPU()
-	batchStart := 0
-
-	if batchCount > len(blockKeys) {
-		batchCount = len(blockKeys)
-	}
-	clients := make([]BlobClient, batchCount)
-	for c := 0; c < batchCount; c++ {
-		client, err := s.blobStore.NewClient(ctx)
-		if err != nil {
-			storeIndex.Dispose()
-			return longtaillib.Longtail_StoreIndex{}, err
-		}
-		clients[c] = client
-	}
-
-	var wg sync.WaitGroup
-
-	for batchStart < len(blockKeys) {
-		batchLength := batchCount
-		if batchStart+batchLength > len(blockKeys) {
-			batchLength = len(blockKeys) - batchStart
-		}
-		batchBlockIndexes := make([]longtaillib.Longtail_BlockIndex, batchLength)
-		wg.Add(batchLength)
-		for batchPos := 0; batchPos < batchLength; batchPos++ {
-			i := batchStart + batchPos
-			blockKey := blockKeys[i]
-			go func(client BlobClient, batchPos int, blockKey string) {
-				storedBlockData, _, err := readBlobWithRetry(
-					ctx,
-					client,
-					blockKey)
-
-				if err != nil {
-					wg.Done()
-					return
-				}
-
-				blockIndex, errno := longtaillib.ReadBlockIndexFromBuffer(storedBlockData)
-				if errno != 0 {
-					wg.Done()
-					return
-				}
-
-				blockPath := GetBlockPath("chunks", blockIndex.GetBlockHash())
-				if blockPath == blockKey {
-					batchBlockIndexes[batchPos] = blockIndex
-				} else {
-					log.Printf("Block %s name does not match content hash, expected name %s\n", blockKey, blockPath)
-				}
-
-				wg.Done()
-			}(clients[batchPos], batchPos, blockKey)
-		}
-		wg.Wait()
-		writeIndex := 0
-		for i, blockIndex := range batchBlockIndexes {
-			if !blockIndex.IsValid() {
-				continue
-			}
-			if i > writeIndex {
-				batchBlockIndexes[writeIndex] = blockIndex
-			}
-			writeIndex++
-		}
-		batchBlockIndexes = batchBlockIndexes[:writeIndex]
-		batchStoreIndex, errno := longtaillib.CreateStoreIndexFromBlocks(batchBlockIndexes)
-		for _, blockIndex := range batchBlockIndexes {
-			blockIndex.Dispose()
-		}
-		if errno != 0 {
-			storeIndex.Dispose()
-			return longtaillib.Longtail_StoreIndex{}, longtaillib.ErrnoToError(errno, longtaillib.ErrENOMEM)
-		}
-		newStoreIndex, errno := longtaillib.MergeStoreIndex(storeIndex, batchStoreIndex)
-		batchStoreIndex.Dispose()
-		storeIndex.Dispose()
-		storeIndex = newStoreIndex
-		//		blockIndexes = append(blockIndexes, batchBlockIndexes[:writeIndex]...)
-		batchStart += batchLength
-		log.Printf("Scanned %d/%d blocks in %s\n", batchStart, len(blockKeys), blobClient.String())
-	}
-
-	for c := 0; c < batchCount; c++ {
-		clients[c].Close()
-	}
-
-	return storeIndex, nil
-}
-
 func buildStoreIndexFromStoreBlocks(
 	ctx context.Context,
 	s *remoteStore,
@@ -614,65 +398,6 @@ func storeIndexWorkerReplyErrorState(
 	}
 }
 
-func readStoreIndexes(
-	ctx context.Context,
-	client BlobClient) (longtaillib.Longtail_StoreIndex, []string, error) {
-
-	var storeIndex longtaillib.Longtail_StoreIndex
-	storeIndexBuffer, _, err := readBlobWithRetry(ctx, client, "store.lsi")
-	if err == nil {
-		storeIndex, _ = longtaillib.ReadStoreIndexFromBuffer(storeIndexBuffer)
-	}
-
-	if !storeIndex.IsValid() {
-		errno := 0
-		storeIndex, errno = longtaillib.CreateStoreIndexFromBlocks([]longtaillib.Longtail_BlockIndex{})
-		if errno != 0 {
-			return longtaillib.Longtail_StoreIndex{}, nil, errors.Wrapf(longtaillib.ErrnoToError(errno, longtaillib.ErrEIO), "readStoreStoreIndex: longtaillib.CreateStoreIndexFromBlocks() failed")
-		}
-	}
-
-	var indexNames []string
-	blobs, err := client.GetObjects("index/")
-	if err != nil {
-		storeIndex.Dispose()
-		return longtaillib.Longtail_StoreIndex{}, nil, err
-	}
-
-	for _, blob := range blobs {
-		if blob.Size == 0 {
-			continue
-		}
-		if strings.HasSuffix(blob.Name, ".plsi") {
-			indexData, _, err := readBlobWithRetry(ctx, client, blob.Name)
-			if err != nil {
-				continue
-			}
-			partialStoreIndex, errno := longtaillib.ReadStoreIndexFromBuffer(indexData)
-			if errno != 0 {
-				continue
-			}
-			storeIndexNew, errno := longtaillib.MergeStoreIndex(storeIndex, partialStoreIndex)
-			partialStoreIndex.Dispose()
-			if errno != 0 {
-				continue
-			}
-			storeIndex.Dispose()
-			storeIndex = storeIndexNew
-			indexNames = append(indexNames, blob.Name)
-		}
-	}
-	return storeIndex, indexNames, nil
-}
-
-func readStoreStoreIndex(
-	ctx context.Context,
-	client BlobClient) (longtaillib.Longtail_StoreIndex, error) {
-
-	storeIndex, _, err := readStoreIndexes(ctx, client)
-	return storeIndex, err
-}
-
 func onPreflighMessage(
 	s *remoteStore,
 	storeIndex longtaillib.Longtail_StoreIndex,
@@ -701,23 +426,6 @@ func onGetExistingContentMessage(
 	message.asyncCompleteAPI.OnComplete(existingContentIndex, 0)
 }
 
-func updateStoreIndex(
-	storeIndex longtaillib.Longtail_StoreIndex,
-	addedBlockIndexes []longtaillib.Longtail_BlockIndex) (longtaillib.Longtail_StoreIndex, error) {
-	addedStoreIndex, errno := longtaillib.CreateStoreIndexFromBlocks(addedBlockIndexes)
-	if errno != 0 {
-		return longtaillib.Longtail_StoreIndex{}, errors.Wrap(longtaillib.ErrnoToError(errno, longtaillib.ErrENOMEM), "contentIndexWorker: longtaillib.CreateStoreIndexFromBlocks() failed")
-	}
-
-	updatedStoreIndex, errno := longtaillib.MergeStoreIndex(addedStoreIndex, storeIndex)
-	addedStoreIndex.Dispose()
-	if errno != 0 {
-		updatedStoreIndex.Dispose()
-		return longtaillib.Longtail_StoreIndex{}, errors.Wrap(longtaillib.ErrnoToError(errno, longtaillib.ErrENOMEM), "contentIndexWorker: longtaillib.MergeStoreIndex() failed")
-	}
-	return updatedStoreIndex, nil
-}
-
 func contentIndexWorker(
 	ctx context.Context,
 	s *remoteStore,
@@ -744,7 +452,7 @@ func contentIndexWorker(
 	if accessType == Init {
 		saveStoreIndex = true
 	} else {
-		storeIndex, err = readStoreStoreIndex(ctx, client)
+		storeIndex, _, err = readStoreIndex(ctx, client)
 		if err != nil {
 			log.Printf("contentIndexWorker: readStoreStoreIndex() failed with %v", err)
 		}
@@ -776,7 +484,7 @@ func contentIndexWorker(
 				return errors.Wrapf(longtaillib.ErrnoToError(errno, longtaillib.ErrENOMEM), "contentIndexWorker: buildStoreIndexFromStoreBlocks() failed")
 			}
 			log.Printf("Rebuilt remote index with %d blocks\n", len(storeIndex.GetBlockHashes()))
-			err := updateRemoteStoreIndex(ctx, client, storeIndex)
+			err := writeStoreIndex(ctx, client, storeIndex)
 			if err != nil {
 				log.Printf("Failed to update store index in store %s\n", s.String())
 				saveStoreIndex = true
@@ -881,20 +589,20 @@ func contentIndexWorker(
 		if errno != 0 {
 			return errors.Wrap(longtaillib.ErrnoToError(errno, longtaillib.ErrENOMEM), s.blobStore.String())
 		}
-		err := updateRemoteStoreIndex(ctx, client, addedStoreIndex)
+		err := writeStoreIndex(ctx, client, addedStoreIndex)
 		if err != nil {
 			log.Printf("Retrying store index in store %s\n", s.String())
-			err = updateRemoteStoreIndex(ctx, client, addedStoreIndex)
+			err = writeStoreIndex(ctx, client, addedStoreIndex)
 		}
 		if err != nil {
 			log.Printf("Retrying 500 ms delayed store index in store %s\n", s.String())
 			time.Sleep(500 * time.Millisecond)
-			err = updateRemoteStoreIndex(ctx, client, addedStoreIndex)
+			err = writeStoreIndex(ctx, client, addedStoreIndex)
 		}
 		if err != nil {
 			log.Printf("Retrying 2 s delayed store index in store %s\n", s.String())
 			time.Sleep(2 * time.Second)
-			err = updateRemoteStoreIndex(ctx, client, addedStoreIndex)
+			err = writeStoreIndex(ctx, client, addedStoreIndex)
 		}
 
 		if err != nil {
